Fix wrong indentation calculation when nesting block elements
When nesting inside another block element, indentation based block
elements need a correctly assigned :indent property.

Currently the leading whitespace is used for this. However, this breaks
down when an indentation based block element is nested e.g. as first
element in a list item:

    * * nested item

The outer list item has a correct indentation of 1 but the inner one has
an incorrect indentation of 1 - this should be 3. The reason is that the
indentation needs to be counted from the left margin in this case.

To achieve this we now track the left margin position (might change due
to blockquote markers) and use that for calucating the needed
indentation.